Split into ChangeCurosrUI and ChangeCursorWorld
